refactor: change config file to be more structured and add transient analysis

+++ src/main/java/org/qesm/app/model/sampling/runner/PerfectSampleRunner.java
@@ -83,7 +83,7 @@ public class PerfectSampleRunner implements SamplerRunner {
         return hist;
     }
 
-    public int getAvgStepsPlusStdDev(int sigmaCount) {
+    public int getAvgStepsPlusStdDev(double sigmaCount) {
         return (int) Math.round(getAvgSteps() + sigmaCount * getStdDevSteps());
     }
